fixed #2 - double run on same host

+++ lib/capistrano/git-submodule-strategy.rb
@@ -35,8 +35,10 @@ class Capistrano::Git
     # make sure the submodules are up-to-date
     # and copy everything to the release path
     def release
-      git :clone, '--depth=1', '--recursive', '-b', fetch(:branch), "file://#{repo_path}", release_path
-      context.execute("find #{release_path} -name '.git*' | xargs -I {} rm -rfv {}")
+      unless context.test(:test, '-e', release_path) && context.test(:ls, '-A', release_path)
+        git :clone, '--depth=1', '--recursive', '-b', fetch(:branch), "file://#{repo_path}", release_path
+        context.execute("find #{release_path} -name '.git*' | xargs -I {} rm -rfv {}")
+      end
     end
   end
 end
